forwarding: extract fabric cross-chassis path into fabric.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the fabric
link resolution / skip classification, fabric-redirect selection (plain
and zone-encoded), the redirect_via_fabric_if_needed / prefer-local
candidate logic, the HA cluster-peer-return fast path, and the shared
kernel-neighbor-state classifier (classify_neighbor_state /
neighbor_state_usable / NeighborStateClass) plus the three fabric/
neighbor diagnostic counters into a new forwarding/fabric.rs submodule.

Functions are moved verbatim; no logic, signature, or hot-path change.
Visibility is preserved: each item's pub(super) becomes
pub(in crate::afxdp) so the mod.rs glob re-export
(pub(in crate::afxdp) use fabric::*) keeps every external call site
(crate::afxdp::forwarding::...) resolving unchanged. record_fabric_skip
stays private next to its only callers. The new submodule uses
super::* like the existing host_inbound submodule.

Validation: cargo build clean (0 errors, same 175 pre-existing
warnings). No #[inline] attribute involved in this batch.
